refactor(card): add domain barrel, route consumers through it
Front-door the card composable domain with an index barrel and migrate
every external consumer off deep module imports. Establishes
useCardMutations as the domain's public card-write seam.
