+++ src/views/audio-reader/term-popover/add-card-panel.vue
@@ -4,7 +4,7 @@ import { useI18n } from 'vue-i18n'
 import { useMemberDecksQuery } from '@/api/decks'
 import { useInsertCardAtMutation } from '@/api/cards'
 import { useLastDeck } from '@/composables/last-deck'
-import { useCardLimitGate } from '@/composables/card/limit-gate'
+import { useCardLimitGate } from '@/composables/card'
 import { useToast } from '@/composables/toast'
 import { emitSfx } from '@/sfx/bus'
 import UiButton from '@/components/ui-kit/button.vue'
