+++ src/views/deck/card-editor/card-face-uploader.vue
@@ -6,8 +6,7 @@ import UiIcon from '@/components/ui-kit/icon.vue'
 import UiTooltip from '@/components/ui-kit/tooltip.vue'
 import FaceImageDropzone from './face-image-dropzone.vue'
 import UiButton from '@/components/ui-kit/button.vue'
-import { cardEditorKey } from '@/composables/card/list-controller'
-import { CARD_IMAGE_MAX_BYTES, useFaceImageUpload } from '@/composables/card/face-image-upload'
+import { cardEditorKey, CARD_IMAGE_MAX_BYTES, useFaceImageUpload } from '@/composables/card'
 import { cardImageUrl } from '@/api/media'
 import { CARD_ATTRIBUTES_DEFAULTS } from '@/utils/deck/defaults'
 import { emitSfx } from '@/sfx/bus'
