+++ src/views/deck/card-editor/list-item-card.vue
@@ -2,8 +2,7 @@
 import CardFaceUploader from './card-face-uploader.vue'
 import { useI18n } from 'vue-i18n'
 import { inject, onMounted, ref, useTemplateRef } from 'vue'
-import { cardEditorKey } from '@/composables/card/list-controller'
-import type { CardWithClientId } from '@/composables/card/virtual-list'
+import { cardEditorKey, type CardWithClientId } from '@/composables/card'
 import textEditor from '@/components/card/text-editor.vue'
 import { emitSfx } from '@/sfx/bus'
 import { useWindowRefocusGuard } from '@/composables/ui/window-refocus-guard'
