+++ src/views/deck/deck-hero/actions.vue
@@ -7,7 +7,7 @@ import { computed, inject } from 'vue'
 import { useI18n } from 'vue-i18n'
 import { useStudyModal } from '@/composables/study-session/study-modal'
 import { useDeckSettingsModal } from '@/composables/deck/settings-modal'
-import { cardEditorKey } from '@/composables/card/list-controller'
+import { cardEditorKey } from '@/composables/card'
 import { deckViewShellKey } from '@/composables/deck/view-shell'
 
 const { deck } = defineProps<{ deck: Deck }>()
