+++ src/views/deck/index.vue
@@ -10,7 +10,7 @@ import CardGridEmpty from './card-grid/empty-state.vue'
 import { preloadDeckModes } from './modes'
 import ScrollBar from '@/components/ui-kit/scroll-bar.vue'
 import { useDeckQuery } from '@/api/decks'
-import { cardEditorKey, useCardListController } from '@/composables/card/list-controller'
+import { cardEditorKey, useCardListController } from '@/composables/card'
 import { deckViewShellKey, useDeckViewShell } from '@/composables/deck/view-shell'
 
 const { id: deck_id } = defineProps<{
