+++ src/views/deck/mode-toolbar/bulk-toolbar.vue
@@ -4,7 +4,7 @@ import CardCount from './card-count.vue'
 import UiButton from '@/components/ui-kit/button.vue'
 import UiTag from '@/components/ui-kit/tag.vue'
 import { useI18n } from 'vue-i18n'
-import { useBulkActions } from '@/composables/card/bulk-actions'
+import { useBulkActions } from '@/composables/card'
 
 const { t } = useI18n()
