+++ src/views/deck/mode-toolbar/card-count.vue
@@ -2,7 +2,7 @@
 import UiTag from '@/components/ui-kit/tag.vue'
 import { useI18n } from 'vue-i18n'
 import { inject } from 'vue'
-import { cardEditorKey } from '@/composables/card/list-controller'
+import { cardEditorKey } from '@/composables/card'
 
 const { t } = useI18n()
